requirements.txt
google-genai
elevenlabs
moviepy
fastapi
uvicorn
python-multipart
aiofiles
requests
pillow
python-dotenv
streamlit
